Slight fix for 32 bit systems

+++ src/coil/coil/RenderObj/Spheres.cpp
@@ -48,28 +48,28 @@ RTSpheres::RTSpheres(size_t N):
   //Work computer test render
   size_t spheres_rendered = 0;
 
-  size_t stage_spheres = std::min(10ul, N - spheres_rendered);
+  size_t stage_spheres = std::min(size_t(10ul), N - spheres_rendered);
   if (stage_spheres)
     {
       _renderDetailLevels.push_back(RTSpheres::SphereDetails(magnet::GL::primatives::Sphere::icosahedron, 2, stage_spheres));
       spheres_rendered += stage_spheres;
     }
 
-  stage_spheres = std::min(1000ul, N - spheres_rendered);
+  stage_spheres = std::min(size_t(1000ul), N - spheres_rendered);
   if (stage_spheres)
     {
       _renderDetailLevels.push_back(RTSpheres::SphereDetails(magnet::GL::primatives::Sphere::icosahedron, 1, stage_spheres));
       spheres_rendered += stage_spheres;
     }
 
-  stage_spheres = std::min(10000ul, N - spheres_rendered);
+  stage_spheres = std::min(size_t(10000ul), N - spheres_rendered);
   if (stage_spheres)
     {
       _renderDetailLevels.push_back(RTSpheres::SphereDetails(magnet::GL::primatives::Sphere::icosahedron, 0, stage_spheres));
       spheres_rendered += stage_spheres;
     }
 
-  stage_spheres = std::min(200000ul, N - spheres_rendered);
+  stage_spheres = std::min(size_t(200000ul), N - spheres_rendered);
   if (stage_spheres)
     {
       _renderDetailLevels.push_back(RTSpheres::SphereDetails(magnet::GL::primatives::Sphere::octahedron, 0, stage_spheres));
